Move net-generic types and utils to gosub_net
