+++ crates/gosub_net/Cargo.toml
@@ -17,9 +17,10 @@ bytes = "1"
 http = "1"
 log = "0.4"
 tracing = "0.1"
+thiserror = "1"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-tokio = { version = "1.52.3", features = ["rt"] }
+tokio = { version = "1.52.3", features = ["rt", "io-util", "sync"] }
 reqwest = { version = "0.13.3", features = ["rustls", "stream", "blocking"] }
