+++ crates/gosub_net/src/net/mod.rs
@@ -1,3 +1,5 @@
 pub mod events;
 pub mod null_emitter;
 pub mod observer;
+pub mod types;
+pub mod utils;
